Update SWT/Eclipse and consume from maven central
The required artifacts have been available on maven central for a few
years now.

Delete the jars in local repository and update maven coordinates to
point to the published artifacts.

The set of dependency versions is based on Eclipse 4.16 (2020-06) which
was the last version to require only Java 8.

This task is a prerequisite for #71.